Fence casework write delegates
