+++ control-plane/tests/test_service_persistence_ingest_case_lifecycle.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import ast
 from dataclasses import replace
 from datetime import timedelta
 import pathlib
@@ -11,6 +12,7 @@ if str(TESTS_ROOT) not in sys.path:
 
 import _service_persistence_support as support
 from _service_persistence_support import ServicePersistenceTestBase
+import aegisops_control_plane.service as service_module
 from aegisops_control_plane.case_workflow import CaseWorkflowService
 from aegisops_control_plane.detection_lifecycle import DetectionIntakeService
 from aegisops_control_plane.evidence_linkage import EvidenceLinkageService
@@ -264,6 +266,37 @@ class IngestCaseLifecyclePersistenceTests(ServicePersistenceTestBase):
             rationale="Delegated disposition.",
             recorded_at=observed_at,
         )
+
+    def test_case_workflow_compatibility_delegates_are_fenced_from_service_class(
+        self,
+    ) -> None:
+        service_source = pathlib.Path(service_module.__file__).read_text(
+            encoding="utf-8",
+        )
+        service_tree = ast.parse(service_source)
+        service_class = next(
+            node
+            for node in service_tree.body
+            if isinstance(node, ast.ClassDef)
+            and node.name == "AegisOpsControlPlaneService"
+        )
+        direct_service_methods = {
+            node.name for node in service_class.body if isinstance(node, ast.FunctionDef)
+        }
+        case_workflow_write_methods = {
+            "record_case_observation",
+            "record_case_lead",
+            "record_case_recommendation",
+            "record_case_handoff",
+            "record_case_disposition",
+        }
+
+        self.assertTrue(
+            case_workflow_write_methods.issubset(
+                set(dir(support.AegisOpsControlPlaneService))
+            )
+        )
+        self.assertFalse(case_workflow_write_methods & direct_service_methods)
 
     def test_service_delegates_detection_intake_and_triage_operations(self) -> None:
         store, _ = support.make_store()
